+++ CHANGELOG.md
@@ -83,6 +83,10 @@
   via the flag `--package`.
   ([Louis Pilfold](https://github.com/lpil))
 
+- The error message when failing to decrypt the local Hex API key is now more
+  informative and helpful.
+  ([Moritz Böhme](https://github.com/MoritzBoehme))
+
 ### Language server
 
 - The language server now supports go-to-definition, find-references and rename